Move common componets to folders and fix imports
