+++ src/components/common/AutomaticAllocationModal/automaticAllocation.accordion.tsx
