+++ src/components/common/AutomaticAllocationModal/automaticAllocation.modal.tsx
