+++ src/pages/buildings.tsx
@@ -4,7 +4,7 @@ import * as C from '@chakra-ui/react';
 import { BsFillPenFill, BsFillTrashFill } from 'react-icons/bs';
 
 import { ColumnDef } from '@tanstack/react-table';
-import Navbar from 'components/common/navbar.component';
+import Navbar from 'components/common/NavBar/navbar.component';
 import BuildingsService from 'services/api/buildings.service';
 import {
   Building,
@@ -12,8 +12,8 @@ import {
   UpdateBuilding,
 } from 'models/common/building.model';
 import RegisterModal from 'components/buildings/register.modal';
-import Dialog from 'components/common/dialog.component';
-import DataTable from 'components/common/dataTable.component';
+import Dialog from 'components/common/Dialog/dialog.component';
+import DataTable from 'components/common/DataTable/dataTable.component';
 import { appContext } from 'context/AppContext';
 import { sortBuildings } from 'utils/sorter';
