+++ src/pages/classrooms.tsx
@@ -16,9 +16,9 @@ import { BsFillPenFill, BsFillTrashFill } from 'react-icons/bs';
 
 import { ColumnDef } from '@tanstack/react-table';
 import RegisterModal from 'components/classrooms/register.modal';
-import DataTable from 'components/common/dataTable.component';
-import Dialog from 'components/common/dialog.component';
-import Navbar from 'components/common/navbar.component';
+import DataTable from 'components/common/DataTable/dataTable.component';
+import Dialog from 'components/common/Dialog/dialog.component';
+import Navbar from 'components/common/NavBar/navbar.component';
 import { appContext } from 'context/AppContext';
 import Classroom from 'models/common/classroom.model';
 import { useContext, useEffect, useState } from 'react';
