+++ src/pages/subjects/subjects.tsx
@@ -1,5 +1,5 @@
-import Navbar from 'components/common/navbar.component';
-import DataTable from 'components/common/dataTable.component';
+import Navbar from 'components/common/NavBar/navbar.component';
+import DataTable from 'components/common/DataTable/dataTable.component';
 import { ColumnDef } from '@tanstack/react-table';
 
 import {
@@ -16,7 +16,7 @@ import {
 import { BsFillPenFill, BsFillTrashFill } from 'react-icons/bs';
 import { useContext, useEffect, useState } from 'react';
 import SubjectRegisterModal from 'components/subjects/subjectRegister.modal';
-import Dialog from 'components/common/dialog.component';
+import Dialog from 'components/common/Dialog/dialog.component';
 import { appContext } from 'context/AppContext';
 import SubjectsService from 'services/api/subjects.service';
 import { sortSubjectsResponse } from 'utils/subjects/subjects.sorter';
@@ -25,7 +25,10 @@ import {
   subjectTypeFormatter,
 } from 'utils/subjects/subjects.formatter';
 import { SubjectsResponseCode } from 'utils/enums/subjects.enum';
-import { CreateSubject, UpdateSubject } from 'models/http/requests/subject.request.models';
+import {
+  CreateSubject,
+  UpdateSubject,
+} from 'models/http/requests/subject.request.models';
 import { SubjectResponse } from 'models/http/responses/subject.response.models';
 
 function Subjects() {
@@ -138,9 +141,9 @@ function Subjects() {
   const [subjects, setSubjects] = useState<Array<SubjectResponse>>([]);
 
   const [isUpdateSubject, setIsUpdateSubject] = useState<boolean>(false);
-  const [selectedSubject, setSelectedSubject] = useState<SubjectResponse | undefined>(
-    undefined,
-  );
+  const [selectedSubject, setSelectedSubject] = useState<
+    SubjectResponse | undefined
+  >(undefined);
 
   const toast = useToast();
   const toastSuccess = (message: string) => {
